Move credential provider to separate package
